requirements.txt
# Core dependencies
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
responses>=0.23.0

# Development tools
black>=23.0.0
flake8>=6.0.0

# Rate limiting (removed; legacy scraping disabled)

# For better HTTP handling
urllib3>=2.0.0

# Modern HTTP client used for BDL and Odds API
httpx>=0.27.0

# Official NBA stats client and dependency
nba_api>=1.3.0
pandas>=2.2.0